+++ src/datascript/impl/entity.cljs
@@ -8,9 +8,8 @@
   {:pre [(satisfies? dc/IDB db)
          (satisfies? dc/ISearch db)
          (satisfies? dc/IIndexAccess db)]}
-  (when eid
-    (when-let [e (dc/resolve-eid db eid)]
-      (Entity. db e false {}))))
+  (when-let [e (dc/entid db eid)]
+    (Entity. db e false {})))
 
 (defn- entity-attr [db a datoms]
   (if (dc/multival? db a)